+++ packages/learn-card-bridge-http/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @learncard/create-http-bridge
 
+## 1.1.199
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.1.5
+
+## 1.1.198
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.1.4
+
 ## 1.1.197
 
 ### Patch Changes