+++ packages/learn-card-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @learncard/cli
 
+## 3.3.138
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.1.5
+
+## 3.3.137
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.1.4
+
 ## 3.3.136
 
 ### Patch Changes