+++ services/learn-card-discord-bot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # learn-card-discord-bot
 
+## 1.1.187
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.1.5
+
+## 1.1.186
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.1.4
+
 ## 1.1.185
 
 ### Patch Changes